+++ src/XrdXrootd/XrdXrootdProtocol.cc
@@ -798,8 +798,8 @@ int XrdXrootdProtocol::StatGen(struct stat &buf, char *xxBuff, int xxLen,
 
 // Format the default response: <devid> <size> <flags> <mtime>
 //
-   m = snprintf(xxBuff, xxLen, "%lld %lld %d %ld",
-                Dev.uuid, fsz, flags, buf.st_mtime);
+   m = snprintf(xxBuff, xxLen, "%lld %lld %d %lld",
+                Dev.uuid, fsz, flags, (long long) buf.st_mtime);
 // if (!xtnd || m >= xxLen) return xxLen;
 //
 
@@ -808,8 +808,9 @@ int XrdXrootdProtocol::StatGen(struct stat &buf, char *xxBuff, int xxLen,
    char *origP = xxBuff;
    char *nullP = xxBuff + m++;
    xxBuff += m; xxLen -= m;
-   n = snprintf(xxBuff, xxLen, "%ld %ld %04o ",
-                buf.st_ctime, buf.st_atime, buf.st_mode&07777);
+   n = snprintf(xxBuff, xxLen, "%lld %lld %04o ",
+                (long long) buf.st_ctime, (long long) buf.st_atime,
+                buf.st_mode&07777);
    if (n >= xxLen) return m;
    xxBuff += n; xxLen -= n;
